Inject zustand store hooks via context
