+++ src/h5web/vis-packs/models.ts
@@ -1,4 +1,4 @@
-import type { ElementType } from 'react';
+import type { ElementType, ReactNode } from 'react';
 import type { IconType } from 'react-icons';
 import type { Entity } from '../providers/models';
 
@@ -6,9 +6,14 @@ export interface VisContainerProps {
   entity: Entity;
 }
 
+export interface ConfigProviderProps {
+  children: ReactNode;
+}
+
 export interface VisDef {
   name: string;
   Icon: IconType;
   Toolbar?: ElementType;
   Container: ElementType<VisContainerProps>;
+  ConfigProvider?: ElementType<ConfigProviderProps>;
 }
